refactor(package): ui와 domain 패키지로 클래스를 구분한다
