+++ src/main/java/bowling/domain/LastFrame.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.domain;
 
 import java.util.LinkedList;
 import java.util.stream.Collectors;
