+++ src/test/java/bowling/LastFrameTest.java
@@ -1,5 +1,6 @@
 package bowling;
 
+import bowling.domain.LastFrame;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
