+++ src/test/java/bowling/NormalFrameTest.java
@@ -1,5 +1,7 @@
 package bowling;
 
+import bowling.domain.Frame;
+import bowling.domain.NormalFrame;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
