+++ src/test/java/bowling/PinsTest.java
@@ -1,5 +1,6 @@
 package bowling;
 
+import bowling.domain.Pins;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
